Spotbugs, up rank to 15 and lower threshold from high to default

+++ docs/src/main/java/org/geotools/tutorial/csv2/CSVWriteTest.java
@@ -86,8 +86,10 @@ public class CSVWriteTest {
     @After
     public void removeTemporaryLocations() throws IOException {
         File list[] = tmp.listFiles();
-        for (int i = 0; i < list.length; i++) {
-            list[i].delete();
+        if (list != null) {
+            for (int i = 0; i < list.length; i++) {
+                list[i].delete();
+            }
         }
         tmp.delete();
     }
